refactor : MatchSection.types.ts로 types 이동
