+++ components/matchCard/MatchCard.types.ts
@@ -4,3 +4,7 @@ export interface MatchCardProps {
   matchId: string;
   nickname: string;
 }
+
+export type QueueTypeMapper = {
+  [index: number]: string;
+};
