+++ components/matchSection/MatchSection.types.ts
@@ -1,0 +1,5 @@
+export type { RecentMatchUserInfo } from 'types';
+
+export interface MatchSectionProps {
+  nickname: string;
+}
